fix(ui): fixed rerender loop in home dashboard
