+++ ui/package.json
@@ -166,5 +166,5 @@
     "test-ui": "vitest --ui --coverage"
   },
   "type": "module",
-  "version": "2.17.0"
+  "version": "2.17.1"
 }
